requirements.txt
anthropic
python-pptx>=0.6.21
